Moved MainActivity from app to ui module
