+++ ui/build.gradle.kts
@@ -31,6 +31,7 @@ dependencies {
 
     implementation(Deps.Kotlin.kotlin_std)
     implementation(Deps.Android.compat)
+    implementation(Deps.Android.constraint_layout)
 
     testImplementation(Deps.Test.junit)
     androidTestImplementation(Deps.Test.runner)
